✨ feat(type_system): Enhance type handling with support for Int8, Int16, UInt8, UInt16, and Never types; update related parsing and type inference logic

+++ src/lexer/token.rs
@@ -18,7 +18,7 @@ pub enum TokenType {
     Operator { operator: OperatorType },
     StringLiteral { value: String },
     CharLiteral { value: char },
-    IntegerLiteral { value: u64 },
+    IntegerLiteral { value: i128 },
     DecimalLiteral { value: f64 },
     BooleanLiteral { value: bool },
     NullLiteral,
